requirements.txt
pydantic==2.10.6
pydantic-ai==0.0.37
requests==2.32.3
python-dotenv==1.0.0
PyGithub==2.1.1
docker==7.0.0
pytest==7.4.3
black==23.11.0
